+++ src/kernel_evm/CHANGES.md
@@ -33,6 +33,7 @@
 - Block hash is no longer in Tx object, receipt and BIP. (!10520)
 - Compute real block hashes. (!10442)
 - EVM evaluation framework (testing for the EVM execution) is now available. (!10221)
+- Compute {state,transactions,receipt}_root in block. (!10545)
 
 ## Version fc06d63568b1be253088ff5d6e422fd80ed3d2c2